Rename AbstractTheiaCloudOperator to BasicTheiaCloudOperator
Also remove the abstract so it can directly be used.
Use it directly in the DefaultTheiaCloudOperatorModule.
Remove the DefaultTheiaCloudOperator as it is no longer needed.
Extract LeaderElection in separate method for easier customization.
Also make the other methods in the leader election protected.
